+++ api/data/src/main/kotlin/data/extend/BufferedImage.kt
@@ -166,7 +166,7 @@ fun BufferedImage.rotate(angle: Double): BufferedImage {
 
 fun BufferedImage.boundBox(xStart: Int, xEnd: Int): Okvir {
     val maxCount = 5
-    val boundBox = Okvir(-1, -1, -1, -1)
+    val boundBox = Okvir(-1.0, -1.0, -1.0, -1.0)
 
     //up to down
     var count = 0
@@ -174,7 +174,7 @@ fun BufferedImage.boundBox(xStart: Int, xEnd: Int): Okvir {
         for (x in xStart until this.width - xEnd) {
             if (!this.getHSV(x, y).is_white()) count++
             if (count > maxCount) {
-                boundBox.y0 = y
+                boundBox.y0 = y.toDouble()
                 break@start
             }
         }
@@ -186,7 +186,7 @@ fun BufferedImage.boundBox(xStart: Int, xEnd: Int): Okvir {
         for (x in xStart until this.width - xEnd) {
             if (!this.getHSV(x, y).is_white()) count++
             if (count > maxCount) {
-                boundBox.y1 = y
+                boundBox.y1 = y.toDouble()
                 break@start
             }
         }
@@ -198,7 +198,7 @@ fun BufferedImage.boundBox(xStart: Int, xEnd: Int): Okvir {
         for (y in 0 until this.height) {
             if (!this.getHSV(x, y).is_white()) count++
             if (count > maxCount) {
-                boundBox.x0 = x
+                boundBox.x0 = x.toDouble()
                 break@start
             }
         }
@@ -210,7 +210,7 @@ fun BufferedImage.boundBox(xStart: Int, xEnd: Int): Okvir {
         for (y in 0 until this.height) {
             if (!this.getHSV(x, y).is_white()) count++
             if (count > maxCount) {
-                boundBox.x1 = x
+                boundBox.x1 = x.toDouble()
                 break@start
             }
         }
